Aggerate calculator (2023) for NUCES Fast
CLI-based program written in C for calculating Aggerate for admission in BS-Programs. The program assumes 1100 Total marks for both F.SC and Matriculation. It takes the current method (2023) of calculation of the aggerate: 10 % Matric + 40 % F.SC + 50 % Fast Entrance Exam.
Note: This program is irrelevant for NTS-based admission as of this update.